requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.5.1
billiard==3.3.0.23
bleach==1.4.3
celery==3.1.24
coveralls==1.1
cryptography==1.7.2
Django==1.10.2
django-autocomplete-light>=2.0.0,<3.0.0
django-constance==1.3.3
django-cors-headers==1.2.2
django-extensions==1.7.6
django-filter==1.0.1
django-grappelli==2.9.1
django-mailbox==4.5.4
django-picklefield==0.3.2
djangorestframework==3.5.3
djangorestframework-jwt==1.8.0
docutils==0.12
drfdocs==0.0.11
elasticsearch==5.0.1
geoip2==2.4.0
geopy==1.11.0
gunicorn==19.6.0
html5lib==0.9999999
httmock==1.2.5
httplib2==0.9.2
jira==1.0.7
kombu==3.0.37
langdetect==1.0.7
maxminddb==1.2.1
oauth2==1.9.0.post1
oauthlib==2.0.0
pika==0.10.0
psycopg2==2.6.2
PyJWT==1.4.2
pymongo==3.3.0
python-dateutil==2.5.3
pytz==2016.7
requests==2.11.1
requests-oauthlib==0.7.0
requests-toolbelt==0.7.0
selenium==3.0.2
six==1.10.0
testfixtures==4.12.0
tlslite==0.4.9
tweepy==3.5.0
urllib3==1.18
